SE/django_project/tennis_app/requirements.txt
asgiref==3.4.1
attrs==21.2.0
boto3==1.20.33
botocore==1.23.33
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
defusedxml==0.7.1
Django==3.0
django-storages==1.12.3
djangorestframework==3.13.1
gunicorn==20.1.0
httpie==2.6.0
idna==3.3
importlib-metadata==4.8.1
iniconfig==1.1.1
jmespath==0.10.0
oauthlib==3.1.1
packaging==21.2
pluggy==1.0.0
psycopg2==2.8.6
py==1.11.0
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.5
python-dateutil==2.8.2
python-magic==0.4.24
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
s3transfer==0.5.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.8
whitenoise==5.3.0
zipp==3.6.0
